chore: rename homepage file to home
